+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 
+- Added customizable switch motor strength and duration (https://github.com/Novakasa/brickrail/issues/129)
 - If any sensor in layout has a color different than None, the train will only react to markers with actually used colors. (https://github.com/Novakasa/brickrail/discussions/127)
 
 ### Fixed